+++ backend/src/main/java/com/birder/bird_observation_project/exceptions/ObservationNotFoundException.java
@@ -1,7 +1,7 @@
 package com.birder.bird_observation_project.exceptions;
 
 public class ObservationNotFoundException extends RuntimeException{
-    public ObservationNotFoundException(Long id){
+    public ObservationNotFoundException(int id){
         super("No observation found with id: "+id+".");
     }
 }
